ath79: implement LOADER_FLASH_ALT for Engenius boards
These boards have an extra "failsafe" image on flash
for recovery when the kernel fails to boot.

They also use the OKLI kernel loader, which originally
would halt when the kernel is missing or fails to boot.

Now, with this added feature and variable,
OKLI kernel loader will load the failsafe image instead of halting,
in case of the real kernel being damaged or missing,
which otherwise would require serial console for recovery.

Signed-off-by: Michael Pratt <mcpratt@pm.me>